---------- Bugtraq Message ---------- fixed #158: Add lazy load for taxonomy tree terms fixed #160: Add URL prefix filter fixed #159: Add modules filter fixed #157: Individual CPT rules are lost when saving widget options -------------------------------------
git-svn-id: https://plugins.svn.wordpress.org/dynamic-widgets@689417 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ trunk/mods/custompost_module.php
@@ -15,7 +15,11 @@
 			parent::admin();
 			self::customPosts();
 			self::customTax();
-			self::customArchive();
+
+			if ( array_key_exists('cp_archive', $GLOBALS['DW']->dwoptions) ) {
+				self::customArchive();
+			}
+
 		}
 
 		public static function customArchive() {
@@ -48,10 +52,16 @@
 
 			// Custom Post Type
 			self::$post_types = get_post_types($args, 'objects', 'and');
+
 			foreach ( self::$post_types as $type => $ctid ) {
+				if (! array_key_exists($type, $GLOBALS['DW']->dwoptions) ) {
+					continue;
+				}
+
 				// Prepare
 				self::$opt = $DW->getDWOpt($widget_id, $type);
 				$tax_list = get_object_taxonomies($type, 'objects');
+				$tax_list = apply_filters('dynwid_taxonomies', $tax_list);
 
 				// Output
 				echo '<input type="hidden" name="post_types[]" value="' . $type . '" />';
@@ -60,7 +70,7 @@
 				echo __('Show widget on', DW_L10N_DOMAIN) . ' ' . $ctid->label . '? ' . ( ($ctid->hierarchical || count($tax_list) > 0) ? '<img src="' . $DW->plugin_url . 'img/info.gif" alt="info" onclick="divToggle(\'custom_' . $type . '\');" />' : '' ) . '<br />';
 				echo '<div>';
 				echo '<div id="custom_' . $type . '" class="infotext">';
-				echo ( $ctid->hierarchical ? '<p>' . DW_Page::infoText() . '</p>' : '' );
+				echo ( $ctid->hierarchical ) ? '<p>' . DW_Page::infoText() . '</p>' : '';
 				echo ( (count($tax_list) > 0) ? '<p>' . __('All exceptions (Titles and Taxonomies) work in a logical OR condition. That means when one of the exceptions is met, the exception rule is applied.', DW_L10N_DOMAIN) . '</p>' : '' );
 				echo '</div>';
 				echo '</div>';
@@ -69,20 +79,7 @@
 				echo '<br />';
 
 				$opt_single = $DW->getDWOpt($widget_id, $type);
-/*
-				$opt_individual = $DW->getDWOpt($_GET['id'], 'individual');
-				if ( $opt_individual->count > 0 ) {
-					$individual = TRUE;
-					$count_individual = '(' . $ctid->label . ' Posts: ' . count($opt_single->act) . ')';
-				}
-				echo '<input type="checkbox" name="individual" value="1" ' . ( (isset($individual) && $individual)  ? 'checked="checked"' : '' ) . ' />';
-				echo ' <label for="individual">' . __('Make exception rule available to individual ' . $ctid->label . '.', DW_L10N_DOMAIN) . ' ' . ( ($opt_individual->count > 0)  ? $count_individual : '' ) . '</label>';
 
-				// Individual posts and tags
-				foreach ( $opt_single->act as $singlepost ) {
-					echo '<input type="hidden" name="' . $type . '_act[]" value="' . $singlepost . '" />';
-				}
-*/
 				// Taxonomy in Custom Post Type
 				foreach ( $tax_list as $tax_type ) {
 					// Prepare
@@ -116,7 +113,14 @@
 						} else {
 							$childs = $opt_tax_childs->act;
 						}
-						self::prtTax($tax_type->name, $tree, $opt_tax->act, $childs, $type . '-tax_' . $tax_type->name);
+
+						echo '<input type="hidden" id="' . $type . '-tax_' . $tax_type->name . '_act" name="' . $type . '-tax_' . $tax_type->name . '_act" value="' . implode(',', $opt_tax->act) . '" />';
+						if ( isset($opt_tax_childs) ) {
+							echo '<input type="hidden" id="' . $type . '-tax_' . $tax_type->name . '_childs_act" name="' . $type . '-tax_' . $tax_type->name . '_childs_act" value="' . implode(',', $opt_tax_childs->act) . '" />';
+						}
+
+						// self::prtTax($tax_type->name, $tree, $opt_tax->act, $childs, $type . '-tax_' . $tax_type->name);
+						self::prtTax($widget_id, $tax_type->name, $tree, $opt_tax->act, $childs, $type . '-tax_' . $tax_type->name);
 						echo '</div>';
 						echo '</div>';
 					}
@@ -128,7 +132,7 @@
 
 		public static function customTax() {
 			$DW = $GLOBALS['DW'];
-			$widget_id = $GLOBALS['widget_id'];			
+			$widget_id = $GLOBALS['widget_id'];
 
 			$args = array(
 				'public'   => TRUE,
@@ -137,9 +141,13 @@
 
 			if ( function_exists('is_tax') ) {
 				$taxlist = get_taxonomies($args, 'objects', 'and');
+				$taxlist = apply_filters('dynwid_taxonomies', $taxlist);
 
 				if ( count($taxlist) > 0 ) {
 					foreach ( $taxlist as $tax_id => $tax ) {
+						if (! array_key_exists('tax_' . $tax_id, $GLOBALS['DW']->dwoptions) ) {
+							continue;
+						}
 
 						// Getting the linked post type : Only Pages and CPT supported
 						$cpt_label = array();
@@ -157,11 +165,6 @@
 									$cpt_label[ ] = self::$post_types[$obj]->label;
 							}
 
-							/* if ( $obj == 'page' ) {
-								$cpt_label[ ] = _('Pages');
-							} else if ( isset(self::$post_types[$obj]) ) {
-								$cpt_label[ ] = self::$post_types[$obj]->label;
-							} */
 						}
 
 						if ( count($cpt_label) > 0 ) {
@@ -190,7 +193,7 @@
 
 							echo '<div>';
 							echo '<div id="custom_' . $ct . '" class="infotext">';
-							echo ( $tax->hierarchical ? '<p>' . DW_Page::infoText() . '</p>' : '' );
+							echo ( $tax->hierarchical ) ? '<p>' . DW_Page::infoText() . '</p>' : '';
 							echo ( (count($t) > 0) ? '<p>' . __('All exceptions work in a logical OR condition. That means when one of the exceptions is met, the exception rule is applied.', DW_L10N_DOMAIN) . '</p>' : '' );
 							echo '</div>';
 							echo '</div>';
@@ -207,7 +210,14 @@
 								} else {
 									$childs = $opt_ct_archive_childs->act;
 								}
-								self::prtTax($tax->name, $tree, $opt_ct_archive->act, $childs, $ct);
+
+								echo '<input type="hidden" id="' . $ct . '_act" name="' . $ct . '_act" value="' . implode(',', $opt_ct_archive->act) . '" />';
+								if ( isset($opt_ct_archive_childs) ) {
+									echo '<input type="hidden" id="' . $ct . '_childs_act" name="' . $ct . '_childs_act" value="' . implode(',', $opt_tax_childs->act) . '" />';
+								}
+
+								self::prtTax($widget_id, $tax->name, $tree, $opt_ct_archive->act, $childs, $ct);
+
 								echo '</div>';
 								echo '</div>';
 							}
@@ -221,8 +231,9 @@
 
 		public static function getTaxChilds($term, $arr, $id, $i) {
 			$tax = get_terms($term, array('hide_empty' => FALSE, 'parent' => $id));
+			return $tax;
 
-			foreach ($tax as $t ) {
+/*			foreach ($tax as $t ) {
 				if (! in_array($t->term_id, $i) && $t->parent == $id ) {
 					$i[ ] = $t->term_id;
 					$arr[$t->term_id] = array();
@@ -231,37 +242,50 @@
 				}
 			}
 
-			return $arr;
+			return $arr; */
 		}
 
-		public static function prtTax($tax, $terms, $terms_act, $terms_childs_act, $prefix) {
+		public static function prtTax($widget_id, $tax, $terms, $terms_act, $terms_childs_act, $prefix) {
 			$DW = &$GLOBALS['DW'];
 
-			foreach ( $terms as $pid => $childs ) {
+			// foreach ( $terms as $pid => $childs ) {
+			foreach ( $terms as $term ) {
 				$run = TRUE;
 
-				if ( $DW->wpml ) {
+/*				if ( $DW->wpml ) {
 					include_once(DW_MODULES . 'wpml_module.php');
 					$wpml_id = DW_WPML::getID($pid, 'tax_' . $tax);
 					if ( $wpml_id > 0 && $wpml_id <> $pid ) {
 						$run = FALSE;
 					}
+				} */
+
+				if ( $DW->wpml ) {
+					include_once(DW_MODULES . 'wpml_module.php');
+
+					$wpml_id = DW_WPML::getID($term->term_id, 'tax_' . $tax);
+					if ( $wpml_id > 0 && $wpml_id <> $term->term_id ) {
+						$run = FALSE;
+					}
 				}
 
 				if ( $run ) {
-					$term = get_term_by('id', $pid, $tax);
+					// $term = get_term_by('id', $pid, $tax);
 
 					echo '<div style="position:relative;left:15px;">';
-					echo '<input type="checkbox" id="' . $prefix . '_act_' . $pid . '" name="' . $prefix . '_act[]" value="' . $pid . '" ' . ( isset($terms_act) && count($terms_act) > 0 && in_array($pid, $terms_act) ? 'checked="checked"' : '' ) . ' onchange="chkChild(\'' . $prefix . '\', ' . $pid . ')" /> <label for="' . $prefix . '_act_' . $pid . '">' . $term->name . '</label><br />';;
+					echo '<input type="checkbox" id="' . $prefix . '_act_' . $term->term_id . '" name="' . $prefix . '_chkbx[]" value="' . $term->term_id . '" ' . ( isset($terms_act) && count($terms_act) > 0 && in_array($term->term_id, $terms_act) ? 'checked="checked"' : '' ) . ' onchange="chkChild(\'' . $prefix . '\', ' . $term->term_id . ')" /> <label for="' . $prefix . '_act_' . $term->term_id . '">' . $term->name . '</label>';
+					echo ( $terms_childs_act !== FALSE ) ? ' <span title=" Click to expand " onclick="term_tree(\'' . $widget_id . '\', \'' . $tax . '\', ' . $term->term_id . ', \'' . $prefix . '\');return false;"><img src="' . $DW->plugin_url .'/img/arrow-down.png" /></span>' : '';
+					echo '<br />';
 
 					if ( $terms_childs_act !== FALSE ) {
-						echo '<div style="position:relative;left:15px;">';
-						echo '<input type="checkbox" id="' . $prefix . '_childs_act_' . $pid . '" name="' . $prefix . '_childs_act[]" value="' . $pid . '" ' . ( isset($terms_childs_act) && count($terms_childs_act) > 0 && in_array($pid, $terms_childs_act) ? 'checked="checked"' : '' ) . ' onchange="chkParent(\'' . $prefix . '\', ' . $pid . ')" /> <label for="' . $prefix . '_childs_act_' . $pid . '"><em>' . __('All childs', DW_L10N_DOMAIN) . '</em></label><br />';
+						echo '<div id="child_' . $prefix . $term->term_id  . '" style="position:relative;left:15px;display:none;">';
+						echo '<input type="checkbox" id="' . $prefix . '_childs_act_' . $term->term_id . '" name="' . $prefix . '_childs_chkbx[]" value="' . $term->term_id . '" ' . ( isset($terms_childs_act) && count($terms_childs_act) > 0 && in_array($term->term_id, $terms_childs_act) ? 'checked="checked"' : '' ) . ' onchange="chkParent(\'' . $prefix . '\', ' . $term->term_id . ')" /> <label for="' . $prefix . '_childs_act_' . $term->term_id . '"><em>' . __('All childs', DW_L10N_DOMAIN) . '</em></label><br />';
+						echo '<div id="tree_' . $prefix . $term->term_id . '"></div>';
 						echo '</div>';
 
-						if ( count($childs) > 0 ) {
+						/* if ( count($childs) > 0 ) {
 							self::prtTax($tax, $childs, $terms_act, $terms_childs_act, $prefix);
-						}
+						} */
 					}
 					echo '</div>';
 				}
@@ -277,12 +301,13 @@
 				'_builtin' => FALSE
 			);
 			$post_types = get_post_types($args, 'objects', 'and');
-			foreach ( $post_types as $ctid ) {
-				$option[key($post_types)] = $ctid->label;
+			foreach ( $post_types as $type => $ctid ) {
+				$option[$type] = $ctid->label;
 			}
 
 			// Adding Custom Taxonomies to $DW->dwoptions
 			$taxonomy = get_taxonomies($args, 'objects', 'and');
+			$taxonomy = apply_filters('dynwid_taxonomies', $taxonomy);
 			foreach ( $taxonomy as $tax_id => $tax ) {
 				$option['tax_' . $tax_id] = $tax->label;
 			}
